requirements.txt
crewai
crewai-tools
langchain
litellm
google-generativeai
python-dotenv
fastapi 
uvicorn